android app: ask permission in the application, not widget
opsie
