+++ android/app/src/main/java/org/toni/customfetch_android/SettingsActivity.kt
@@ -1,7 +1,12 @@
 package org.toni.customfetch_android
 
+import android.app.AlertDialog
+import android.content.Intent
 import android.content.res.AssetManager
+import android.os.Build
 import android.os.Bundle
+import android.os.Environment
+import android.provider.Settings
 import android.util.Log
 import androidx.appcompat.app.AppCompatActivity
 import androidx.preference.PreferenceFragmentCompat
@@ -22,6 +27,25 @@ class SettingsActivity : AppCompatActivity() {
                 .commit()
         }
         supportActionBar?.setDisplayHomeAsUpEnabled(true)
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
+            if (!Environment.isExternalStorageManager()) {
+                AlertDialog.Builder(this)
+                    .setTitle("Grant external storage management permission")
+                    .setMessage("Customfetch needs permissions to manage external storage for writing config files\n"+
+                            "By default we going to read/write the following directories:\n"+
+                            "/storage/emulated/0/.config/\n"+
+                            "/storage/emulated/0/.config/customfetch/")
+                    // The dialog is automatically dismissed when a dialog button is clicked.
+
+                    .setPositiveButton("Grant permission"
+                    ) { _, _ ->
+                        val intent = Intent(Settings.ACTION_MANAGE_ALL_FILES_ACCESS_PERMISSION)
+                        startActivity(intent)
+                    }
+                    .setIcon(android.R.drawable.ic_dialog_alert)
+                    .show()
+            }
+        }
     }
 
     class SettingsFragment : PreferenceFragmentCompat() {
